services/ingest/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
redis==5.0.1
httpx==0.26.0
tenacity==8.2.3
pytest==8.0.0
pytest-asyncio==0.23.3
